Added a new material struct which is now used as ImGui TextureId.
This allows for different shader programs to be used in draw calls for image widgets.
Used this system to create a font and video material, where the video material has 3 textures for it's YUV channels.
For now the YUV channels are 1:1 mapped to RGB channels. Proper YUV->RGB conversion is up next.