Added better frame skipping for videos that cannot be decoded fast enough.
The video no longer hangs on a key frame that is skipped through. Instead it keeps
decoding and presenting P/B-frames until the I-frame is near enough to skip to it.
To figure if the next key frame is near a enough to skip to it just in time we keep
track of the lowest decode time of all previously decoded key frames. Other than
that we don't keep the time we need to decode audio frames into account.
Because of both that the lowest key frame decode time is used and because
audio frame decoding times are ignored, the key frame is usually still coming
in a little too late.
Future improvements may be to offload audio decoding to it's own thread.