Library entries can now be dragged from an Explorer window in to a Sequentia window.
Library links no longer have a relative path entry. An old project path should be enough to find the entries again.
All strings that are stored in actions are now copied. That way we can safely delete them when actions are overwritten after some undos, followed by a new action.
Changed the library UI to a table, for now I plan to only support files, not folders in the library to keep things simple.