Converted all C-style enums to C++-style enum classes.
I got into trouble having same enum item names (compile error). Same names are allowed on C++-style enums.
The only exception so far is the SeqDialogOption enum, which is used as a bitmask. Bitmasking is not possible
with a strongly typed enum without bitwise operator overloading, which I'm not a fan of, but who knows, might
happen in the future still if it will end up to be the lesser evil.