SeqUIVideo now plays the last selected video using SeqTaskDecodeVideo and SeqRenderer's
video frame write functions.
This brought an issue to light, which is that AVFormatContext seems to store playback state.
Because we have a single AVFormatContext in SeqLibraryLink videos will continue playing
where they left of last time. Which is almost never what we want, even more so because we
want to be able to play the same video at different playback times. So AVFormatContext
need to be removed from SeqLibraryLink and be generated when needed for playback.