Allowed for clips to be dragged between channels. This required some big structure changes.
When dragging a clip between channels we need to know where the clip came from for SeqAction (Undo/Redo).
Now a SeqClip is no longer dragged around, but a new SeqClipProxy is used instead.
A SeqClipProxy has a pointer to a SeqClip but it's own location data, now expressed in the new SeqClipLocation component.
During the dragging the original SeqClip is hidden while the SeqClipProxy is been drawn.
After the SeqClipProxy is (mouse button)released an action is added/executed on SeqProject to actually move the SeqClip to the new SeqClipProxy location.
- Made the SeqActionFactory functions less verbose (yeah we get it, the action factory creates actions).
- Stashed away some thoughts in the TODO.txt file.