Implemented a basic SeqPlayer. SeqUIVideo is now displaying SeqPlayer output instead of
SeqDecoder output.
A SeqPlayer is what is composing the output of a SeqScene. It will spin up and keep track of
SeqTaskDecodeVideo objects, and compose the outputs of those decoders together into the
final edited video.
This is now also used to show the preview scene which is updated every time a new video is
selected in the SeqUILibrary.
SeqProject now has a width and height (now hardcoded set to 1920x1080) which will be used by
the SeqPlayer compositor as coordinate space to place clips. Furthermore it's the size of the
frame buffer the compositor is rendering it's output to.
Changed the renderer so we can change between the back buffer for ImGui rendering and a
frame buffer for SeqPlayer rendering. The glViewport is also set, and a different projection
matrix is applied based on if the material is used for SeqPlayer compositing or ImGui rendering.
- SeqDecoder now uses avformat_seek_file isntead of av_seek_frame. This allows for searching
  just before the requested time instead of just after. This fixes a bunch of ugly seek behavior
  in the previous version of the decoder where the Video would hang for a while until the playtime
  caught up with the video frames buffer.
- Updated SeqTime.h with a bunch of parenthesis so everything still works if an expression is used
  wit the macro instead of a straight number.