Fixed a bug where often the previous preview video was shown. This happened because
SeqClipPlayer objects got invalid by accidental 'clip' field memory overwrites when a another
video was selected in SeqUILibrary. This resulted in SeqClipPlayer objects with a clip definition
of clip A with a decoder for clip B etc.
Fixed that by changing the preview scene update code to use SeqActions that are not recorded in
the undo/redo list.
Executing actions emit events which are used in SeqPlayer to properly dispose of SeqClipPlayers
so invalid ones are not used.
- Changed when action events are fired from after the action execution to before it. This seems
 to work in all cases the events are used right now but maybe something more advanced Is
 needed in the future.
- Added a DoAndForgetAction public function in SeqProject which does an action but does
not add it in the undo/redo list. Also the action will action's data memory will be freed afterwards.
- Restructed clipPlayer getting in SeqPlayer to make it more flexible. Moved SeqClipPlayer disposal
to a separte function so we're not doing the same thing in 3 locations.