Fixed a bug in SeqDecoder where a stream's time_base was ignored. Futhermore SeqDecoder
now uses dts instead of pts for getting the timing of a packet or frame.
Millisecond timebase was working for all test videos until now quite by accident. Testing
another video (mpeg4 instead of hevc) which had a timebase of frame time instead of 1/1000
seconds the error was evident. Now SeqDecoder takes Sequentia time (as defined in SeqTime)
and converts it to the right stream time based on the stream's time base. Vice versa packet and
frame times are now converted back to SeqTime in SeqDecoder getters.
The new test video also showed that packet pts and frame pts was not as reliable as packet dts
and frame pkt_dts as time indicator. This video had a pts of 0 on every other frame and a pts
increasing by 2 on the frame in between. dts however had expected results of n + 1 per frame.
This causeed only half of the frames to be displayed of the new test video, with the pts to dts
update all frames are shown.