Added a SeqClipGroup. This groups clips that can be played by the same decoder together.
The most typical use case is a video with a video and audio stream which are often used together.
Now a grouped set of clips will use only one SeqTaskDecodeVideo task in SeqPlayer.
The grouping is done for videos in the preview scene.
There are still a number of open issues mentioned in the updated TODO.txt