- Renderer overhaul -
Added OpenGL message callback which brought a number of issues to light which are fixed here:
The main issue was that there was only one Vertex Array Object (VAO) used, on which
glEnableVertexAttribArray and glVertexAttribPointer is called multiple times, which doesn't make
much sense. Really we need a VAO per SeqMaterial as each material can contain a different set of
vertex attributes per vertex shader type. So now the program handle and associated variables are
moved from SeqMaterialInstance to SeqMaterial. Each SeqMaterial has it's own VAO which is
bound the same time when textures are bound on a SeqMaterialInstance.
Other changes:
- Shader compilation now show errors when there are any.
- Fixed OpenGL issues for when there is no uv or color attribute in a shader.
- Removed a bunch of over protective OpenGL state backup and restores in SeqRenderer.
- Added some utility functions for manipulating projection matrices.
- Added utility functions for Material and Texture creation.
- Made SeqViewport in SeqRenderer more generic.